+++ src/Microsoft.SqlTools.ServiceLayer/SqlContext/SqlToolsSettings.cs
@@ -143,140 +143,4 @@ namespace Microsoft.SqlTools.ServiceLayer.SqlContext
             }
         }
     }
-
-    public interface ISqlToolsSettingsValues
-    {
-        IntelliSenseSettings IntelliSense { get; set; }
-        QueryExecutionSettings QueryExecutionSettings { get; set; }
-        FormatterSettings Format { get; set; }
-        ObjectExplorerSettings ObjectExplorer { get; set; }
-    }
-
-    public class CompoundToolsSettingsValues: ISqlToolsSettingsValues
-    {
-        private List<ISqlToolsSettingsValues> priorityList = new List<ISqlToolsSettingsValues>();
-        private SqlToolsSettingsValues defaultValues;
-        public CompoundToolsSettingsValues(ISqlToolsSettingsValues mssql, ISqlToolsSettingsValues all)
-        {
-            Validate.IsNotNull(nameof(mssql), mssql);
-            Validate.IsNotNull(nameof(all), all);
-            priorityList.Add(mssql);
-            priorityList.Add(all);
-            // Always add in a fallback which has default values to be used.
-            defaultValues = new SqlToolsSettingsValues(createDefaults: true);
-            priorityList.Add(defaultValues);
-        }
-
-        private T GetSettingOrDefault<T>(Func<ISqlToolsSettingsValues, T> lookup)
-            where T : new()
-        {
-            T value = priorityList.Select( (settings) => lookup(settings)).Where(val => val != null).FirstOrDefault();
-            return value != null ? value : new T();
-        }
-
-        /// <summary>
-        /// Gets or sets the detailed IntelliSense settings
-        /// </summary>
-        public IntelliSenseSettings IntelliSense
-        { 
-            get
-            {
-                return GetSettingOrDefault((settings) => settings.IntelliSense);
-            } 
-            set
-            {
-                priorityList[0].IntelliSense = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the query execution settings
-        /// </summary>
-        [JsonProperty("query")]
-        public QueryExecutionSettings QueryExecutionSettings
-        { 
-            get
-            {
-                return GetSettingOrDefault((settings) => settings.QueryExecutionSettings);
-            } 
-            set
-            {
-                priorityList[0].QueryExecutionSettings = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the formatter settings
-        /// </summary>
-        [JsonProperty("format")]
-        public FormatterSettings Format
-        { 
-            get
-            {
-                return GetSettingOrDefault((settings) => settings.Format);
-            } 
-            set
-            {
-                priorityList[0].Format = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the formatter settings
-        /// </summary>
-        [JsonProperty("objectExplorer")]
-        public ObjectExplorerSettings ObjectExplorer
-        { 
-            get
-            {
-                return GetSettingOrDefault((settings) => settings.ObjectExplorer);
-            } 
-            set
-            {
-                priorityList[0].ObjectExplorer = value;
-            }
-        }
-
-    }
-    /// <summary>
-    /// Class that is used to serialize and deserialize SQL Tools settings
-    /// </summary>
-    public class SqlToolsSettingsValues : ISqlToolsSettingsValues
-    {
-        /// <summary>
-        /// Initializes the Sql Tools settings values
-        /// </summary>
-        public SqlToolsSettingsValues(bool createDefaults = true)
-        {
-            if (createDefaults)
-            {
-                IntelliSense = new IntelliSenseSettings();
-                QueryExecutionSettings = new QueryExecutionSettings();
-                Format = new FormatterSettings();
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the detailed IntelliSense settings
-        /// </summary>
-        public IntelliSenseSettings IntelliSense { get; set; }
-
-        /// <summary>
-        /// Gets or sets the query execution settings
-        /// </summary>
-        [JsonProperty("query")]
-        public QueryExecutionSettings QueryExecutionSettings { get; set; }
-
-        /// <summary>
-        /// Gets or sets the formatter settings
-        /// </summary>
-        [JsonProperty("format")]
-        public FormatterSettings Format { get; set; }
-
-        /// <summary>
-        /// Gets or sets the formatter settings
-        /// </summary>
-        [JsonProperty("objectExplorer")]
-        public ObjectExplorerSettings ObjectExplorer { get; set; }
-    }
 }
